pyproject.toml
[tool.poetry]
name = "gfa"
version = "0.1.0dev"
description = "A library and actor to control the Guide, Focus, and Acquisition cameras."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/gfa"
repository = "https://github.com/sdss/gfa"
documentation = "https://gfa.readthedocs.io/en/latest/"
keywords = ["astronomy", "software"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "gfa", from = "python" }
]
include = ["README.md", "python/gfa/etc/*"]

[tool.poetry.dependencies]
python = "^3.7"
pygments = "^2.2.0"
setuptools = "^42.0.0"
toml = "^0.10.0"
"ruamel.yaml" = "^0.15.61"

# Docs extras. Need to be added to dependencies and not to dev-dependencies because otherwise
# they won't be installed on pip install .[docs]
# See https://github.com/sdispater/poetry/pull/606#issuecomment-437943927
# Need to do poetry install -E docs to actually install everything.
Sphinx = {version = "^2.1", optional = true}
semantic-version = {version = "2.8.0", optional = true}

[tool.poetry.dev-dependencies]
ipython = "^7.9.0"
matplotlib = "^3.1.1"
invoke = "^1.3.0"
flake8 = "^3.7.9"
doc8 = "^0.8.0"
pytest = "^5.2.2"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.8.1"
pytest-sugar = "^0.9.2"
isort = "^4.3.21"
codecov = "^2.0.15"
coverage = {version = "5.0", extras = ["toml"]}

[tool.poetry.extras]
docs = ["Sphinx", "semantic-version"]

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "gfa"
# multi_line_output = false
balanced_wrapping = true
include_trailing_comma = false
length_sort_stdlib = false
lines_after_imports = 2
use_parentheses = true

[tool.coverage.run]
branch = true
include = ["python/gfa/*"]
omit = [
    "*/core/*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[tool.flake8]
ignore = [
    "H101",
    "N",
    "W504",
]
max-line-length = 99

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.0.0"]
